+++ repository.tf
@@ -77,9 +77,12 @@ resource "azuredevops_git_repository_file" "locals" {
 }
 
 resource "azuredevops_git_repository_file" "terraform" {
-  repository_id       = azuredevops_git_repository.landing_zone.id
-  file                = "terraform.tf"
-  content             = templatefile("${path.module}/templates/terraform.tftpl", {})
+  repository_id = azuredevops_git_repository.landing_zone.id
+  file          = "terraform.tf"
+  content = templatefile("${path.module}/templates/terraform.tftpl", {
+    azurerm_version = var.azurerm_version
+    azapi_version   = var.azapi_version
+  })
   branch              = "refs/heads/${azuredevops_git_repository_branch.init.name}"
   commit_message      = "Add terraform.tf"
   overwrite_on_create = true
@@ -93,7 +96,8 @@ resource "azuredevops_git_repository_file" "tags" {
   repository_id = azuredevops_git_repository.landing_zone.id
   file          = "tags.tf"
   content = templatefile("${path.module}/templates/tags.tftpl", {
-    tags = local.tags
+    module_version_tags = var.module_version_tags
+    tags                = local.tags
   })
   branch              = "refs/heads/${azuredevops_git_repository_branch.init.name}"
   commit_message      = "Add tags.tf"